Chains

NNHelferlein.NeuNet — Type

abstract type NeuNet end

Mother type for DNN hierarchy with implementation for a chain of layers.

Signatures:

(n::NeuNet)(x) = (for l in n.layers; x = l(x); end; x)
(m::NeuNet)(d::Knet.Data) = mean( m(x,y) for (x,y) in d)

NNHelferlein.Classifier — Type

struct Classifier <: NeuNet

Classifyer with nll loss.

Signatures:

(m::Classifier)(x,y) = nll(m(x), y)

NNHelferlein.Regressor — Type

struct Regressor

Regression network with square loss.

Signatures:

(m::Regression)(x,y) = sum(abs2.( m(x) - y))

Layers

Missing docstring.

Missing docstring for Layer. Check Documenter's build log for details.

NNHelferlein.Dense — Type

struct Dense  <: Layer

Default Dense layer.

Constructors:

Dense(w, b, actf): default constructor

Dense(i::Int, j::Int; actf=sigm): layer of j neurons with i inputs.

Dense(h5::HDF5File, group::String; trainable=false, actf=sigm): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

NNHelferlein.Conv — Type

struct Conv  <: Layer

Default Conv layer.

Constructors:

Conv(w, b, padding, actf): default constructor

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu, padding=0): layer with o kernels of size (w1,w2) for an input of i layers.

Conv(h5::HDF5File, group::String; trainable=false, actf=relu): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

NNHelferlein.Pool — Type

struct Pool <: Layer

Pooling layer.

Constructors:

Pool(): default 2×2 pooling

Pool(k...): user-defined pooling

NNHelferlein.Flat — Type

struct Flat <: Layer

Default flatten layer.

Constructors:

Flat(): with no options.

NNHelferlein.PyFlat — Type

struct PyFlat <: Layer

Flatten layer with optional Python-stype flattening (row-major). This layer can be used if pre-trained weight matrices from tensorflow are applied after the flatten layer.

Constructors:

PyFlat(; python=true): if true, row-major flatten is performed.

NNHelferlein.Embed — Type

struct Embed

Simple type for an embedding layer to embed a virtual onehot-vector into a smaller number of neurons by linear combination. The onehot-vector is virtual, because not the vector, but only the index of the "one" in the vector has to be provided as Integer value (or a minibatch of integers).

Fields:

w

actf

Constructors:

Embed(i,j; actf=identity) = new(param(j,i), actf): with input size i, output size j and default activation function idendity.

Signatures:

(l::Embed)(x) = l.actf.(w[:, permutedims(hcat(x...))]) default embedding of input vector x.

NNHelferlein.Predictions — Type

struct Predictions

Simple wrapper around a Dense layer without activation function that can used as output layer (because all loss-functions of the package assume raw output activations).

Constructors:

Predictions(i::Int, j:Int): with input size i, output size j activation function idendity.

Predictions(h5::HDF5File, group::String; trainable=false): with an hdf5-object and group name of the output layer.

Missing docstring.

Missing docstring for Softmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Layer. Check Documenter's build log for details.

Data providers

NNHelferlein.DataLoader — Type

abstract type DataLoader

Mother type for minimatch iterators.

Tabular data

Tabular data is normally provided in table form (csv, ods) row-wise, i.e. one sample per row. The helper functions can read the tables and generate Knet compatible iterators of minibatches.

NNHelferlein.dataframe_read — Function

dataframe_read(fname)

Read a data table from an ODS- or CSV-file with one sample per row and return a DataFrame with the data.

NNHelferlein.dataframe_minibatches — Function

dataframe_minibatches(data::DataFrames.DataFrame; size=256, ignore=[], teaching=:y, regression=true)

Make Knet-conform minibatches from a dataframe with one sample per row.

ignore: defines a list of column names to be ignored

teaching: defines the column with teaching input. Default is ":y"

regression: if true, the teaching input is interpreted as scalar value (converted to Float32); otherwise teaching input is used as class labels (converted to UInt8).

NNHelferlein.dataframe_split — Function

function dataframe_split(df::DataFrames.DataFrame; teaching=:y, fr=0.2, balanced=true)

Split data, organised row-wise in a DataFrame into train and valid sets.

Arguments:

df: data

teaching: name or index of column with teaching input (y)

fr: fraction of data to be used for validation

balanced: if true, result datasets will be balanced by oversampling. Returned datasets will be bigger as expected but include the same numbers of samples for each class.

Image data

Images as data should be provided in directories with the directory names denoting the class labels. The helpers read from the root of a directory tree in which the first level of sub-dirs tell the class label. All images in the tree under a class label are read as instances of the respective class. The following tree will generate the classes daisy, rose and tulip:

image_dir/
├── daisy
│   ├── 01
│   │   ├── 01
│   │   ├── 02
│   │   └── 03
│   ├── 02
│   │   ├── 01
│   │   └── 02
│   └── others
├── rose
│   ├── big
│   └── small
└── tulip

NNHelferlein.ImageLoader — Type

struct ImageLoader <: DataLoader
    dir
    i_paths
    i_classes
    classes
    batchsize
    shuffle
    train
    aug_pipl
    pre_proc
end

Iterable image loader.

NNHelferlein.mk_image_minibatch — Function

function mk_image_minibatch(dir, batchsize; split=false, fr=0.2,
                            balanced=false, shuffle=true, train=true,
                            aug_pipl=nothing, pre_proc=nothing)

Return an iterable image-loader-object that provides minibatches of path-names of image files, relative to dir.

Arguments:

dir: base-directory of the image dataset. The first level of sub-dirs are used as class names.

batchsize: size of minibatches

Keyword arguments:

split: return two iterators for training and validation

fr: split fraction

balanced: return balanced data (i.e. same number of instances for all classes). Balancing is achieved via oversampling

shuffle: if true, shuffle the images everytime the iterator restarts

train: if true, minibatches with (x,y) Tuples are provided, if false only x (for prediction)

pre_load: if true all images are loaded in advance; otherwise images are loaded on demand durng training. (option is not implemented yet!)

aug_pipl: augmentation pipeline for Augmentor.jl. Augmentation is performed before the pre_proc-function is applied

pre_proc: function with preprocessing and augmentation algoritms of type x = f(x). In contrast to the augmentation that modifies images, is pre_proc working on Arrays{Float32}.

NNHelferlein.get_class_labels — Function

function get_class_labels(d::DataLoader)

Extracts a list of class labels from a DataLoader.

Training

NNHelferlein.tb_train! — Function

function tb_train!(mdl, opti, trn, vld; epochs=1,
                  lr_decay=0.0, lrd_freq=1, l2=0.0,
                  eval_size=0.2, eval_freq=1,
                  mb_loss_freq=100,
                  cp_freq=1, cp_dir="checkpoints",
                  tb_dir="logs", tb_name="run",
                  tb_text="""Description of tb_train!() run.""",
                  args...)

Train function with TensorBoard integration. TB logs are written with the TensorBoardLogger.jl package. The model is updated (in-place) and the trained model is returned.

Arguments:

mdl: model; i.e. forward-function for the net

opti: Knet-stype optimiser iterator

trn: training data; iterator to provide (x,y)-tuples with minibatches

vld: validation data; iterator to provide (x,y)-tuples with minibatches.

Keyword arguments:

Optimiser:

epochs=1: number of epochs to train

lr_decay=0.0: Leraning rate decay: factor (<1) to reduce the lr.

lrd_freq=1: frequency of learning rate decay steps. Default is to modify the lr after every epoch

l2=0.0: L2 regularisation; implemented as weight decay per parameter

args...: optional keyword arguments for the optimiser can be specified (i.e. lr, gamma, ...).

Model evaluation:

eval_size=0.2: fraction of validation data to be used for calculating loss and accuracy for train and validation data during training.

eval_freq=1: frequency of evaluation; default=1 means evaluation is calculated after each epoch. With eval_freq=10 eveluation is calculated 10 times per epoch.

mb_loss_freq=100: frequency of training loss reporting. default=100 means that 100 loss-values per epoch will be logged to TensorBoard. If mblossfreq is greater then the number of minibatches, loss is logged for each minibatch.

cp_freq=1: frequency of model checkpoints written to disk.

after all other args. Default is to write the model after each epoch with name model.

cp_dir="checkpoints": directory for checkpoints

TensorBoard:

TensorBoard log-directory is created from 3 parts: tb_dir/tb_name/<current date time>.

tb_dir="logs": root directory for tensorborad logs.

tb_name="run": name of training run. tb_name will be used as directory name and should not include whitespace

tb_text: description to be included in the TensorBoard log as text log.

Evaluation

NNHelferlein.predict — Function

function predict(mdl, x; softmax=false)

Return the prediction for x.

Arguments:

mdl: executable network model x: tensor, minibatch or iterator providing minibatches of input data softmax: if true and if model is a ::Classifier the prediction softmax probabilities are rezrned instead of raw activations.

NNHelferlein.predict_top5 — Function

function predict_top5(mdl, x; top_n=5, classes=nothing)

Run the model mdl for data in x and prints the top 5 predictions as softmax probabilities.

Arguments:

top_n: print top n hits instead of 5 classes may be a list of human readable class labels.

ImageNet tools

NNHelferlein.preproc_imagenet — Function

function preproc_imagenet(img)

Image preprocessing for pre-trained ImageNet examples. Preprocessing includes

bring RGB colour values into a range 0-255

standardise of colour values by substracting mean colour values (103.939, 116.779, 123.68) from RGB

changing colour channel sequence from RGB to BGR

Resize is not done, because this may be part of the augmentation pipeline.

Examples:

The function can be used with the image loader; for prediction with a trained model as:

pipl = CropRatio(ratio=1.0) |> Resize(224,224)
images = mk_image_minibatch("./example_pics", 16;
                    shuffle=false, train=false,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

And for training something like:

pipl = Either(1=>FlipX(), 1=>FlipY(), 2=>NoOp()) |>
       Rotate(-5:5) |>
       ShearX(-5:5) * ShearY(-5:5) |>
       RCropSize(224,224)

dtrn, dvld = mk_image_minibatch("./example_pics", 16;
                    split=true, fr=0.2, balanced=false,
                    shuffle=true, train=true,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

Missing docstring.

Missing docstring for precdict_imagenet. Check Documenter's build log for details.