API doc of all exported functions are listed here:

Chains

NNHelferlein.DNN — Type

abstract type DNN end

Mother type for DNN hierarchy with implementation for a chain of layers.

Signatures:

(m::DNN)(x) = (for l in m.layers; x = l(x); end; x)
(m::DNN)(x,y) = m(x,y)
(m::DNN)(d::Knet.Data) = mean( m(x,y) for (x,y) in d)
(m::DNN)(d::Tuple) = mean( m(x,y) for (x,y) in d)
(m::DNN)(d::NNHelferlein.DataLoader) = mean( m(x,y) for (x,y) in d)

NNHelferlein.print_network — Function

function print_network(mdl::DNN)

Print a network summary of any model of Type DNN. If the model has a field layers, the summary of all included layers will be printed recursively.

NNHelferlein.Classifier — Type

struct Classifier <: DNN

Classifyer with nll loss.

Signatures:

(m::Classifier)(x,y) = nll(m(x), y)

NNHelferlein.Regressor — Type

struct Regressor

Regression network with square loss.

Signatures:

(m::Regression)(x,y) = sum(abs2.( m(x) - y))

NNHelferlein.Chain — Type

struct Chain

Simple wrapper to chain layers and execute them one after another.

NNHelferlein.VAE — Type

struct VAE

Type for a generic variational autoencoder.

Constructor:

VAE(encoder, decoder)

Separate predefinded chains (ideally, but not necessarily of type Chain) for encoder and decoder must be specified. The VAE needs the 2 parameters mean and variance to define the distribution of each code-neuron in the bottleneck-layer. In consequence the encoder outputmust be 2 times the size of the decoder input (in case of dense layers: if encoder output is a 8-value vector, 4 codes are defined and the decoder input is a 4-value vector; in case of convolutional layers the number of encoder output channels must be 2 times the number of the encoder input channels - see the examples).

Signatures:

(vae::VAE)(x)
(vae::VAE)(x,y)

Called with one argument, predict will be executed; with two arguments (args x and y should be identical for the autoencoder) the loss will be returned.

Details:

The loss is calculated as the sum of element-wise error squares plus the Kullback-Leibler-Divergence to adapt the distributions of the bottleneck codes:

\[\mathcal{L} = \frac{1}{2} \sum_{i=1}^{n_{outputs}} (t_{i}-o_{i})^{2} - \frac{1}{2} \sum_{j=1}^{n_{codes}}(1 + ln\sigma_{c_j}^{2}-\mu_{c_j}^{2}-\sigma_{c_j}^{2}) \]

Output of the autoencoder is cropped to the size of input before loss calculation (and before prediction); i.e. the output has always the same dimensions as the input, even if the last layer generates a bigger shape.

Layers

NNHelferlein.Layer — Type

abstract type Layer end

Mother type for layers hierarchy.

Fully connected layers

NNHelferlein.Dense — Type

struct Dense  <: Layer

Default Dense layer.

Constructors:

Dense(w, b, actf): default constructor

Dense(i::Int, j::Int; actf=sigm): layer of j neurons with i inputs.

Dense(h5::HDF5.File, group::String; trainable=false, actf=sigm):

Dense(h5::HDF5.File, kernel::String, bias::String; trainable=false, actf=sigm): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

NNHelferlein.Linear — Type

struct Linear  <: Layer

Almost standard dense layer, but functionality inspired by the TensorFlow-layer:

capable to work with input tensors of any number of dimensions

default activation function indetity

optionally without biases.

The shape of the input tensor is preserved; only the size of the first dim is changed from in to out.

Constructors:

Linear(i::Int, j::Int; bias=true, actf=identity) weher i is fan-in and j is fan-out.

Keyword arguments:

bias=true: if false biases are fixed to 0.0

actf=identity: activation function.

NNHelferlein.Embed — Type

struct Embed <: Layer

Simple type for an embedding layer to embed a virtual onehot-vector into a smaller number of neurons by linear combination. The onehot-vector is virtual, because not the vector, but only the index of the "one" in the vector has to be provided as Integer value (or a minibatch of integers).

Fields:

w

actf

Constructors:

Embed(v,d; actf=identity): with vocab size v, embedding depth d and default activation function idendity.

Signatures:

(l::Embed)(x) = l.actf.(w[:,x]) default embedding of input tensor x.

Convolutional

NNHelferlein.Conv — Type

struct Conv  <: Layer

Default Conv layer.

Constructors:

Conv(w, b, padding, actf): default constructor

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu; kwargs...): layer with o kernels of size (w1,w2) for an input of i layers.

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu):

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

Keyword arguments:

padding=0: the number of extra zeros implicitly concatenated at the start and end of each dimension.

stride=1: the number of elements to slide to reach the next filtering window.

dilation=1: dilation factor for each dimension.

... See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function conv4() are supported.

NNHelferlein.DeConv — Type

struct DeConv  <: Layer

Default deconvolution layer.

Constructors:

DeConv(w, b, actf, kwargs...): default constructor

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu, kwargs...): layer with o kernels of size (w1,w2) for an input of i channels.

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu):

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

Keyword arguments:

padding=0: the number of extra zeros implicitly concatenated at the start and end of each dimension (applied to the output).

stride=1: the number of elements to slide to reach the next filtering window (applied to the output).

... See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function deconv4() are supported.

NNHelferlein.Pool — Type

struct Pool <: Layer

Pooling layer.

Constructors:

Pool(;kwargs...): max pooling; without kwargs, 2x2-pooling is performed.

Keyword arguments:

window=2: pooling window size (same for both directions)

...: See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function pool are supported.

NNHelferlein.UnPool — Type

struct UnPool <: Layer

Unpooling layer.

Constructors:

UnPool(;kwargs...): user-defined unpooling

Recurrent

NNHelferlein.RSeqClassifier — Type

struct RSeqClassifier <: Layer

One layer RNN sequence classifyer that works with minimatches of (time) series data. minibatch can be a 2- or 3-dimensional Array. If 2-d, inputs for one step are in one column and the Array has as many colums as steps. If 3-d, the last dimension iterates the samples of the minibatch.

Result is always a 2-d matrix with the output of the units of the last step in each column and one column per sample of the minibatch.

Constructors:

RSeqClassifer(n_inputs::Int, n_units::Int; u_type=:lstm, o...): with number of inputs, number of units and unit type. Internally the type Knet.RNN is used and all keyword arguments of Knet.RNN may be provided.

NNHelferlein.RSeqTagger — Type

struct RSeqTagger <: Layer

One layer RNN sequence classifyer that works with minimatches of (time) series data. minibatch can be a 2- or 3-dimensional Array. If 2-d, inputs for one step are in one column and the Array has as many colums as steps. If 3-d, the last dimension iterates the samples of the minibatch.

Result is an array matrix with the output of the units of all steps for all smaples of the minibatch (with model depth as first and samples of the minimatch as last dimension).

Constructors:

RSeqTagger(n_inputs::Int, n_units::Int; u_type=:lstm, o...): with number of inputs, number of units and unit type. Internally the type Knet.RNN is used and all keyword arguments of Knet.RNN may be provided.

NNHelferlein.hidden_states — Function

function hidden_states(l::<RNN_Type>)

Return the hidden states of one or more layers of an RNN. <RNN_Type> is one of RSeqClassifier, RSeqTagger, Knet.RNN.

NNHelferlein.cell_states — Function

function cell_states(l::<RNN_Type>)

Return the cell states of one or more layers of an RNN only if it is a LSTM. <RNN_Type> is one of RSeqClassifier, RSeqTagger, Knet.RNN.

Others

NNHelferlein.Flat — Type

struct Flat <: Layer

Default flatten layer.

Constructors:

Flat(): with no options.

NNHelferlein.PyFlat — Type

struct PyFlat <: Layer

Flatten layer with optional Python-stype flattening (row-major). This layer can be used if pre-trained weight matrices from tensorflow are applied after the flatten layer.

Constructors:

PyFlat(; python=true): if true, row-major flatten is performed.

NNHelferlein.Softmax — Type

struct Softmax <: Layer

Simple softmax layer to compute softmax probabilities.

Constructors:

Softmax()

NNHelferlein.Dropout — Type

struct Dropout <: Layer

Dropout layer. Implemented with help of Knet's dropout() function that evaluates AutoGrad.recording() to detect if in training or inprediction. Dropouts are applied only if prediction.

Constructors:

Dropout(p) with the dropout rate p.

NNHelferlein.BatchNorm — Type

struct BatchNorm <: Layer

Batchnormalisation layer. Implemented with help of Knet's batchnorm() function that evaluates AutoGrad.recording() to detect if in training or in prediction. In training the moments are updated to record the running averages; in prediction the moments are applied, but not modified.

In addition, optional trainable factor a and bias b are applied:

\[y = a \cdot \frac{(x - \mu)}{(\sigma + \epsilon)} + b\]

Constructors:

Batchnom(; trainable=false, channels=0) will initialise the moments with Knet.bnmoments() and trainable parameters a and b only if trainable==true (in this case, the number of channels must be defined - for CNNs this is the number of feature maps).

Details:

2d, 4d and 5d inputs are supported. Mean and variance are computed over dimensions (2), (1,2,4) and (1,2,3,5) for 2d, 4d and 5d arrays, respectively.

If trainable=true and channels != 0, trainable parameters a and b will be initialised for each channel.

If trainable=true and channels == 0 (i.e. Batchnom(trainable=true)), the params a and b are not initialised by the constructor. Instead, the number of channels is inferred when the first minibatch is normalised as: 2d: size(x)[1] 4d: size(x)[3] 5d: size(x)[4] or 0 otherwise.

NNHelferlein.LayerNorm — Type

struct LayerNorm  <: Layer

Simple layer normalisation (inspired by TFs LayerNormalization). Implementation is from Deniz Yuret's answer to feature request 429 (https://github.com/denizyuret/Knet.jl/issues/492).

The layer performs a normalisation within each sample, not batchwise. Normalisation is modified by two trainable parameters a and b (variance and mean) added to every value of the sample vector.

Constructors:

LayertNorm(depth; eps=1e-6): depth is the number of activations for one sample of the layer.

Signatures:

function (l::LayerNorm)(x; dims=1): normalise x along the given dimensions. The size of the specified dimension must fit with the initialised depth.

Attention Mechanisms

NNHelferlein.AttentionMechanism — Type

abstract type AttentionMechanism

Attention mechanisms follow the same interface and common signatures.

If possible, the algorithm allows precomputing of the projections of the context vector generated by the encoder in a encoder-decoder-architecture (i.e. in case of an RNN encoder the accumulated encoder hidden states).

By default attention scores are scaled according to Vaswani et al., 2017 (Vaswani et al., Attention Is All You Need, CoRR, 2017).

All algorithms use soft attention.

Constructors:

Attn*Mechanism*(dec_units, enc_units; scale=true)
Attn*Mechanism*(units; scale=true)

The one-argument version can be used, if encoder dimensions and decoder dimensions are the same.

Common Signatures:

function (attn::AttentionMechanism)(h_t, h_enc; reset=false)
function (attn::AttentionMechanism)(; reset=false)

Arguments:

h_t: decoder hidden state. If $h_t$ is a vector, its length equals the number of decoder units. If it is a matrix, $h_t$ includes the states for a minibatch of samples and has the size [units, mb].

h_enc: encoder hidden states, 2d or 3d. If $h_{enc}$ is a matrix [units, steps] with the hidden states of all encoder steps. If 3d: [units, mb, steps] encoder states for all minibatches.

reset=false: If the keyword argument is set to true, projections of the encoder states are computed. By default projections are stored in the object and reused until the object is resetted. For attention mechanisms that don't allow precomputation the argument is ignored.

The short form (::AttentionMechanism)(reset=true) can be used to reset the precomputed projections.

Return values

All functions return c and α where α is a matrix of size [mb,steps] with the attention factors for each step and minibatch. c is a matrix of size [units, mb] with the context vector for each sample of the minibatch, calculated as the α-weighted sum of all encoder hidden states $h_{enc}$ for each minibatch.

Attention Mechanisms:

All attention mechanisms calculate attention factors α from scores derived from projections of the encoder hidden states:

\[\alpha = \mathrm{softmax}(\mathrm{score}(h_{enc},h_{t}) \cdot 1/\sqrt{n}))\]

Attention mechanisms implemented:

NNHelferlein.AttnBahdanau — Type

mutable struct AttnBahdanau <: AttentionMechanism

Bahdanau-style (additive, concat) attention mechanism according to the paper:

D. Bahdanau, KH. Co, Y. Bengio, Neural Machine Translation by jointlylearning to align and translate, ICLR, 2015.

\[\mathrm{score}(h_{t},h_{enc}) = v_{a}^{\top}\cdot\tanh(W[h_{t},h_{enc}])\]

Constructors:

AttnBahdanau(dec_units, enc_units; scale=true)
AttnBahdanau(units; scale=true)

NNHelferlein.AttnLuong — Type

mutable struct AttnLuong <: AttentionMechanism

Luong-style (multiplicative) attention mechanism according to the paper (referred as General-type attention): M.-T. Luong, H. Pham, C.D. Manning, Effective Approaches to Attention-based Neural Machine Translation, CoRR, 2015.

\[\mathrm{score}(h_{t},h_{enc}) = h_{t}^{\top} W h_{enc}\]

Constructors:

AttnLuong(dec_units, enc_units; scale=true)
AttnLuong(units; scale=true)

NNHelferlein.AttnDot — Type

mutable struct AttnDot <: AttentionMechanism

Dot-product attention (without trainable parameters) according to the Luong, et al. (2015) paper.

$\mathrm{score}(h_{t},h_{enc}) = h_{t}^{\top} h_{enc}$

Constructors:

AttnDot(; scale=true)

NNHelferlein.AttnLocation — Type

mutable struct AttnLocation <: AttentionMechanism

Location-based attention that only depends on the current decoder state $h_t$ and not on the encoder states, according to the Luong, et al. (2015) paper.

$\mathrm{score}(h_{t}) = W h_{t}$

Constructors:

AttnLocation(len, dec_units; scale=true)

len: maximum sequence length of the encoder to be considered for attention. If the actual length of $h_{enc}$ is bigger as the length of α, attention factors for the remaining states are set to 0.0. If the actual length of h_enc is smaller than α, only the matching attention factors are applied.

dec_units: number of decoder units.

NNHelferlein.AttnInFeed — Type

mutable struct AttnInFeed <: AttentionMechanism

Input-feeding attention that depends on the current decoder state $h_t$ and the next input to the decoder $i_{t+1}$, according to the Luong, et al. (2015) paper.

Infeed attention provides a semantic attention that depends on the next input token.

$\mathrm{score}(h_{t}, i_{t+1}) = W_h h_{t} + W_i i_{t+1} = W [h_t, i_{t+1}]$

Constructors:

AttnInFeed(len, dec_units, fan_in; scale=true)

len: maximum sequence length of the encoder to be considered for attention. If the actual length of $h_{enc}$ is bigger as the length of α, attention factors for the remaining states are set to 0.0. If the actual length of h_enc is smaller than α, only the matching attention factors are applied.

dec_units: number of decoder units.

fan_in: size of the decoder input.

Signature:

function (attn::AttnInFeed)(h_t, inp, h_enc)

h_t: decoder hidden state. If $h_t$ is a vector, its length equals the number of decoder units. If it is a matrix, $h_t$ includes the states for a minibatch of samples and has the size [units, mb].

inp: next decoder input $i_{t+1}$ (e.g. next embedded token of sequence)

h_enc: encoder hidden states, 2d or 3d. If $h_{enc}$ is a matrix [units, steps] with the hidden states of all encoder steps. If 3d: [units, mb, steps] encoder states for all minibatches.

Layers for transformers

NNHelferlein.PositionalEncoding — Type

struct PositionalEncoding <: Layer

Positional encoding layer. Only sincos-style (according to Vaswani, et al., NIPS 2017) is implemented.

The layer takes an array of any any number of dimensions (>=2), calculates the Vaswani-2017-style positional encoding and adds the encoding to each plane of the array.

Data providers

NNHelferlein.DataLoader — Type

abstract type DataLoader

Mother type for minibatch iterators.

Tabular data

Tabular data is normally provided in table form (csv, ods) row-wise, i.e. one sample per row. The helper functions can read the tables and generate Knet compatible iterators of minibatches.

NNHelferlein.dataframe_read — Function

dataframe_read(fname)

Read a data table from an CSV-file with one sample per row and return a DataFrame with the data. (ODS-support is removed because of PyCall compatibility issues of the OdsIO package).

NNHelferlein.dataframe_minibatches — Function

dataframe_minibatches(data::DataFrames.DataFrame; size=256, ignore=[], teaching="y", o...)

Make Knet-conform minibatches of type Knet.data from a dataframe with one sample per row.

Arguments:

ignore: defines a list of column names to be ignored

teaching="y": defines the column name with teaching input. Default is "y". teaching is handled differently, depending on its type: If Int, the teaching input is interpreted as class ids and directly used for training (this assumes that the values range from 1..n). If type is a String, values are interpreted as class labels and convertet to numeric class IDs by calling mk_class_ids(). The list of valid lables and their order can be created by calling mk_class_ids(data.y)[2]. If teaching is a scalar value, regression context is assumed, and the value is used unchanged for training.

other keyword arguments: all keyword arguments accepted by Knet.minibatch() may be used.

NNHelferlein.dataframe_split — Function

function dataframe_split(df::DataFrames.DataFrame;
                         teaching="y", fr=0.2, balanced=true)

Split data, organised row-wise in a DataFrame into train and valid sets.

Arguments:

df: data

teaching="y": name or index of column with teaching input (y)

fr=0.2: fraction of data to be used for validation

shuffle=true: shuffle the rows of the dataframe.

balanced=true: if true, result datasets will be balanced by oversampling. Returned datasets will be bigger as expected but include the same numbers of samples for each class.

NNHelferlein.mk_class_ids — Function

function mk_class_ids(labels)

Take a list with n class labels for n instances and return a list of n class-IDs (of type Int) and an array of lables with the array index of each label corresponds its ID.

Arguments:

labels: List of labels (typically Strings)

Result values:

array of class-IDs in the same order as the input

array of unique class-IDs ordered by their ID.

Examples:

julia> labels = ["blue", "red", "red", "red", "green", "blue", "blue"]
7-element Array{String,1}:
 "blue"
 "red"
 "red"
 "red"
 "green"
 "blue"
 "blue"

julia> mk_class_ids(labels)[1]
7-element Array{Int64,1}:
 1
 3
 3
 3
 2
 1
 1

 julia> mk_class_ids(labels)[2]
3-element Array{String,1}:
 "blue"
 "green"
 "red"

NNHelferlein.MBNoiser — Type

type MBNoiser

Iterator to wrap any Knet.Data iterator of minibatches in order to add random noise. Each value will be multiplied with a random value form Gaussian noise with mean=1.0 and sd=sigma.

Construtors:

MBNoiser(mbs::Knet.Data, σ=1.0)

mbs: iteraor with minibatches

σ: standard deviation for the Gaussian noise

Example:

trn = minibatch(x)
tb_train!(mdl, Adam, MBNoiser(trn, σ=0.1))

Image data

Images as data should be provided in directories with the directory names denoting the class labels. The helpers read from the root of a directory tree in which the first level of sub-dirs tell the class label. All images in the tree under a class label are read as instances of the respective class. The following tree will generate the classes daisy, rose and tulip:

image_dir/
├── daisy
│   ├── 01
│   │   ├── 01
│   │   ├── 02
│   │   └── 03
│   ├── 02
│   │   ├── 01
│   │   └── 02
│   └── others
├── rose
│   ├── big
│   └── small
└── tulip

NNHelferlein.ImageLoader — Type

struct ImageLoader <: DataLoader
    dir
    i_paths
    i_classes
    classes
    batchsize
    shuffle
    train
    aug_pipl
    pre_proc
    pre_load
    i_images
end

Iterable image loader to provide minibatches of images as 4-d-arrays (x,y,rgb,mb).

NNHelferlein.mk_image_minibatch — Function

function mk_image_minibatch(dir, batchsize; split=false, fr=0.2,
                            balanced=false, shuffle=true, train=true,
                            pre_load=false,
                            aug_pipl=nothing, pre_proc=nothing)

Return one or two iterable image-loader-objects that provides minibatches of images. For training each minibatch is a tupel (x,y) with x: 4-d-array with the minibatch of data and y: vector of class IDs as Int.

Arguments:

dir: base-directory of the image dataset. The first level of sub-dirs are used as class names.

batchsize: size of minibatches

Keyword arguments:

split: return two iterators for training and validation

fr: split fraction

balanced: return balanced data (i.e. same number of instances for all classes). Balancing is achieved via oversampling

shuffle: if true, shuffle the images everytime the iterator restarts

train: if true, minibatches with (x,y) Tuples are provided, if false only x (for prediction)

pre_load: if true all images are loaded in advance; otherwise images are loaded on demand durng training. (option is not implemented yet!)

aug_pipl: augmentation pipeline for Augmentor.jl. Augmentation is performed before the pre_proc-function is applied

pre_proc: function with preprocessing and augmentation algoritms of type x = f(x). In contrast to the augmentation that modifies images, is pre_proc working on Arrays{Float32}.

pre_load=false: read all images from disk once when populating the loader (requires loads of memory, but speeds up training).

NNHelferlein.get_class_labels — Function

function get_class_labels(d::DataLoader)

Extracts a list of class labels from a DataLoader.

NNHelferlein.image2array — Function

function image2array(img)

Take an image and return a 3d-array for RGB and a 2d-array for grayscale images with the colour channels as last dimension.

NNHelferlein.array2image — Function

function array2image(arr)

Take a 3d-array with colour channels as last dimension or a 2d-array and return an array of RGB or of Gray as Image.

NNHelferlein.array2RGB — Function

function array2RGB(arr)

Take a 3d-array with colour channels as last dimension or a 2d-array and return always an array of RGB as Image.

Text data

NNHelferlein.WordTokenizer — Type

mutable struct WordTokenizer
    len
    w2i
    i2w
end

Create a word-based vocabulary: every unique word of a String or a list of Strings is assigned to a unique number. The created object includes a list of words (i2w, ordered by their numbers) and a dictionary w2i with the words as keys.

Constructor:

function WordTokenizer(texts; len=nothing, add_ctls=true)

With arguments:

texts: AbstractArray or iterable collection of AbstractArrays to be analysed.

len=nothing: maximum number of different words in the vocabulary. Additional words in texts will be encoded as unknown. If nothing, all words of the texts are included.

add_ctls=true: if true, control words are added in front of the vocabulary (extending the maximum length by 4): "<start>"=>1, "<end>"=>2, "<pad>"=>3 and "<unknown>"=>4.

Signatures:

function (t::WordTokenizer)(w::T; split_words=false, add_ctls=false)
                            where {T <: AbstractString}

Encode a word and return the corresponding number in the vocabulary or the highest number (i.e. "<unknown>") if the word is not in the vocabulary.

The encode-signature accepts the keyword arguments split_words and add_ctls. If split_words==true, the input is treated as a sentence and splitted into single words and an array of integer with the encoded sequence is returned. If add_ctls==true the sequence will be framed by <start> and <end> tokens.

function (t::WordTokenizer)(i::Int)

Decode a word by returning the word corresponding to i or "<unknown>" if the number is out of range of the vocabulary.

function (t::WordTokenizer)(s::AbstractArray{T}; add_ctls=false)
                           where {T <: AbstractString}

Called with an Array of Strings the tokeniser splits the strings into words and returns an Array of Array{Int} with each of the input strings represented by a sequence of Integer values.

function (t::WordTokenizer)(seq::AbstractArray{T}; add_ctls=false)
                                 where {T <: Int}

Called with an Array of Integer values a single string is returned with the decoded token-IDs as words (space-separated).

Base Signatures:

    function length(t::WordTokenizer)

Return the length of the vocab.

Examples:

julia> vocab = WordTokenizer(["I love Julia", "They love Python"]);
Julia> vocab(8)
"Julia"

julia> vocab("love")
5

julia> vocab.(split("I love Julia"))
3-element Array{Int64,1}:
 5
 6
 8

julia> vocab.i2w
9-element Array{String,1}:
 "<start>"
 "<end>"
 "<pad>"
 "<unknown>"
 "love"
 "I"
 "They"
 "Julia"
 "Python"

julia> vocab.w2i
Dict{String,Int64} with 9 entries:
  "I"         => 6
  "<end>"     => 2
  "<pad>"     => 3
  "They"      => 7
  "Julia"     => 8
  "love"      => 5
  "Python"    => 9
  "<start>"   => 1
  "<unknown>" => 4

julia> vocab.([7,5,8])
3-element Array{String,1}:
 "They"
 "love"
 "Julia

julia> vocab.("I love Scala", split_words=true)
3-element Array{Int64,1}:
 6
 5
 4

julia> vocab.([6,5,4])
3-element Array{String,1}:
 "I"
 "love"
 "<unknown>"

julia> vocab("I love Python", split_words=true, add_ctls=true)
5-element Array{Int64,1}:
 1
 6
 5
 9
 2

julia> vocab(["They love Julia", "I love Julia"])
2-element Array{Array{Int64,1},1}:
 [7, 5, 8]
 [6, 5, 8]

NNHelferlein.get_tatoeba_corpus — Function

function get_tatoeba_corpus(lang; force=false,
            url="https://www.manythings.org/anki/")

Download and read a bilingual text corpus from Tatoeba (privided) by ManyThings (https://www.manythings.org). All corpi are English-Language-pairs with different size and quality. Considerable languages include:

fra: French-English, 180 000 sentences

deu: German-English, 227 000 sentences

heb: Hebrew-English, 126 000 sentences

por: Portuguese-English, 170 000 sentences

tur: Turkish-English, 514 000 sentences

The function returns two lists with corresponding sentences in both languages. Sentences are are not processed/normalised/cleaned, but exactly as provided by Tatoeba.

The data is stored in the package directory and only downloaded once.

Arguments:

lang: languagecode

force=false: if true, the corpus is downloaded even if a data file is already saved.

url: base url of ManyThings.

NNHelferlein.seq_minibatch — Function

function seq_minibatch(x, [y,] batchsize; seq_len=nothing, pad=3, o...)

Return an iterator of type Knet.Data with sequence minibatches from a list of sequences. all keyword args of Knet.minibatch() can be used.

All sequences in x are brought to the same length by truncating (if too long) or padding with the token provided as pad.

If y is defined, the minibatches include the sequences for x and training targets y, given as n-dimensional array (as for Knet.minibach()). For sequence-2-sequence minibatches the function seq2seq_minibatch() must be used.

Arguments:

x: An iterable object of sequences.

y: vector or array with training targets

batchsize: size of minibatches

seq_len=nothing: demanded length of sequences in the minibatches. If nothing, all sequences are padded to match with the longest sequence.

pad=3: token, used for padding. The default (3) is the token set by the WordRTokenizer. The token must be compatible with the type of the sequence elements.

o...: any other keyword arguments of Knet.minibatch(), such as shuffle=true or partial=true can be provided.

NNHelferlein.seq2seq_minibatch — Function

function seq2seq_minibatch(x, y, batchsize; seq_len=nothing,
            pad_x=3, pad_y=3, o...)

Return an iterator of type Knet.Data with (x,y) sequence minibatches from two lists of sequences. all keyword args of Knet.minibatch() can be used.

All sequences in x and y are brought to the same length by truncating (if too long) or padding with the token provided as pad.

Arguments:

x: An iterable object of sequences.

y: An iterable object of target sequences.

batchsize: size of minibatches

seq_len=nothing: demanded length of sequences in the minibatches. If nothing, all sequences are padded to match with the longest sequence.

pad_x=3,

pad_y=3: token, used for padding. The token must be compatible with the type of the sequence elements.

o...: any other keyword arguments of Knet.minibatch(), such as shuffle=true or partial=true can be provided.

Training

NNHelferlein.tb_train! — Function

function tb_train!(mdl, opti, trn, vld=nothing; epochs=1,
                  lr_decay=nothing, lrd_steps=5, lrd_linear=false,
                  l2=0.0,
                  eval_size=0.2, eval_freq=1,
                  acc_fun=nothing,
                  mb_loss_freq=100,
                  cp_freq=nothing, cp_dir="checkpoints",
                  tb_dir="logs", tb_name="run",
                  tb_text="""Description of tb_train!() run.""",
                  opti_args...)

Train function with TensorBoard integration. TB logs are written with the TensorBoardLogger.jl package. The model is updated (in-place) and the trained model is returned.

Arguments:

mdl: model; i.e. forward-function for the net

opti: Knet-stype optimiser type

trn: training data; iterator to provide (x,y)-tuples with minibatches

vld: validation data; iterator to provide (x,y)-tuples with minibatches. Set to nothing, if not defined.

Keyword arguments:

Optimiser:

epochs=1: number of epochs to train

lr_decay=nothing: do a leraning rate decay if not nothing: the value given is the final learning rate after lrd_steps steps of decay (lr_decay may be bigger than lr; in this case the leraning rate is increased). lr_decay is only applied if both start learning rate lr and final learning rate lr_decay are defined explicitly. Example: lr=0.01, lr_decay=0.001 will reduce the lr from 0.01 to 0.001 during the training (by default in 5 steps).

lrd_steps=5: number of learning rate decay steps. Default is to modify the lr 5 times during the training.

lrd_linear=false: type of learning rate decay; If false, lr is modified by a constant factor (e.g. 0.9) resulting in an exponential decay. If true, lr is modified by the same step size, i.e. linearly.

l2=0.0: L2 regularisation; implemented as weight decay per parameter

opti_args...: optional keyword arguments for the optimiser can be specified (i.e. lr, gamma, ...).

Model evaluation:

eval_size=0.2: fraction of validation data to be used for calculating loss and accuracy for train and validation data during training.

eval_freq=1: frequency of evaluation; default=1 means evaluation is calculated after each epoch. With eval_freq=10 eveluation is calculated 10 times per epoch.

acc_fun=nothing: function to calculate accuracy. The function must implement the following signature: fun(model; data) where data is an iterator that provides (x,y)-tuples of minibatches. For classification tasks, accuracy from the Knet package is a good choice. For regression a correlation or mean error may be preferred.

mb_loss_freq=100: frequency of training loss reporting. default=100 means that 100 loss-values per epoch will be logged to TensorBoard. If mblossfreq is greater then the number of minibatches, loss is logged for each minibatch.

cp_freq=nothing: frequency of model checkpoints written to disk. Default is nothing, i.e. no checkpoints are written. To write the model after each epoch with name model use freq=1; to write every 2 epochs freq=0.5.

cp_dir="checkpoints": directory for checkpoints

TensorBoard:

TensorBoard log-directory is created from 3 parts: tb_dir/tb_name/<current date time>.

tb_dir="logs": root directory for tensorborad logs.

tb_name="run": name of training run. tb_name will be used as directory name and should not include whitespace

tb_text: description to be included in the TensorBoard log as text log.

Evaluation

NNHelferlein.predict — Function

function predict(mdl, x; softmax=false)

Return the prediction for x.

Arguments:

mdl: executable network model

x: iterator providing minibatches of input data

softmax: if true or if model is of type Classifier the predicted softmax probabilities are returned instead of raw activations.

NNHelferlein.predict_top5 — Function

function predict_top5(mdl, x; top_n=5, classes=nothing)

Run the model mdl for data in x and print the top 5 predictions as softmax probabilities.

Arguments:

top_n: print top n hits

classes: optional list of human readable class labels.

NNHelferlein.hamming_dist — Function

function hamming_dist(p, t; accuracy=false, 
                            ignore_ctls=false, vocab=nothing, 
                            start=nothing, stop=nothing, pad=nothing, unk=nothing)

function hamming_acc(p, t; o...)

function hamming_acc(mdl; data=data, o...)

Return the Hamming distance between two sequences or two minibatches of sequences. Predicted sequences p and teaching input sequences t may be of different length but the number of sequences in the minibatch must be the same.

Arguments:

p, t: n-dimensional arrays of type Int with predictions and teaching input for a minibatch of sequences. Shape of the arrays must be identical except of the first dimension (i.e. the sequence length) that may differ between p and t.

accuracy=false: if false, the mean Hamming distance in the minibatch is returned (i.e. the average number of differences in the sequences). If true, the accuracy is returned for all not padded positions in a range (0.0 - 1.0).

ignore_ctls=false: a vocab is used to replace all '<start>, <end>, <unknwon>, <pad>' tokens by <pad>. If true, padding and other control tokens are treated as normal codes and are not ignored.

vocab=nothing: target laguage vocabulary of type NNHelferlein.WordTokenizer. If defined, the padding token of vocab is used to mask all control tokens in the sequences (i.e. '<start>, <end>, <unknwon>, <pad>').

start, stop, pad, unk: may be used to define individual control tokens. default is nothing.

Details:

The function hamming_acc() is a shortcut to return the accuracy instead of the distance. The signature hamming_acc(mdl; data=data; o...) is for compatibility with acc functions called by train.

ImageNet tools

NNHelferlein.preproc_imagenet — Function

function preproc_imagenet(img)

Image preprocessing for pre-trained ImageNet examples. Preprocessing includes

bring RGB colour values into a range 0-255

standardise of colour values by substracting mean colour values (103.939, 116.779, 123.68) from RGB

changing colour channel sequence from RGB to BGR

Resize is not done, because this may be part of the augmentation pipeline.

Examples:

The function can be used with the image loader; for prediction with a trained model as:

pipl = CropRatio(ratio=1.0) |> Resize(224,224)
images = mk_image_minibatch("./example_pics", 16;
                    shuffle=false, train=false,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

And for training something like:

pipl = Either(1=>FlipX(), 1=>FlipY(), 2=>NoOp()) |>
       Rotate(-5:5) |>
       ShearX(-5:5) * ShearY(-5:5) |>
       RCropSize(224,224)

dtrn, dvld = mk_image_minibatch("./example_pics", 16;
                    split=true, fr=0.2, balanced=false,
                    shuffle=true, train=true,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

NNHelferlein.predict_imagenet — Function

function predict_imagenet(mdl, x; top_n=5)

Predict the ImageNet-class of images from the predefined list of class labels.

NNHelferlein.get_imagenet_classes — Function

function get_imagenet_classes()

Return a list of all 1000 ImageNet class labels.

Other utils

Utils for transformers

NNHelferlein.positional_encoding_sincos — Function

function positional_encoding_sincos(n_embed, n_seq)

Calculate and return a matrix of size [n_embed, n_seq] of positional encoding values following the sin and cos style in the paper Vaswani, A. et al.; Attention Is All You Need; 31st Conference on Neural Information Processing Systems (NIPS 2017), Long Beach, CA, USA, 2017.

NNHelferlein.mk_padding_mask — Function

function mk_padding_mask(x; pad=3)

Make a padding mask; i.e. return an Array of type KnetArray{Float32} (or Array{Float32}) similar to x but with two additional dimension of size 1 in teh middle (this will represent the 2nd seq_len and the number of heads) in multi-head attention and the value 1.0 at each position where x is pad and 0.0 otherwise.

The function can be used for creating padding masks for attention mechanisms.

NNHelferlein.mk_peek_ahead_mask — Function

function mk_peek_ahead_mask(x; dim=1)

Return a matrix of size [n_seq, n_seq] filled with 1.0 and the uppper triangle set to 0.0. Type is KnetArray{Float32} in GPU context, Array{Float32} otherwise. The matrix can be used as peek-ahead mask in transformers.

dim=1 specifies the dimension in which the sequence length is represented. For un-embedded data this is normally 1, i.e. the shape of x is [nseq, nmb]. After embedding the shape probably is [depth, nseq, nmb].

NNHelferlein.dot_prod_attn — Function

function dot_prod_attn(q, k, v; mask=nothing)

Generic scaled dot product attention following the paper of Vaswani et al., (2017), Attention Is All You Need.

Arguments:

q: query of size [depth, n_seq_q, ...]

k: key of size [depth, n_seq_v, ...]

v: value of size [depth, n_seq_v, ...]

mask: mask for attention factors may have different shapes but must be broadcastable for addition to the scores tensor (which as the same size as alpha [n_seq_v, n_seq_q, ...]). In transformer context typical masks are one of: padding mask of size [n_seq_v, ...] or a peek-ahead mask of size [n_seq_v, n_seq_v] (which is only possible in case of self-attention when all seqencee lengths are identical).

q, k, v must have matching leading dimensions (i.e. same depth or embedding). k and v must have the same sequence length.

Return values:

c: context as alpha-weighted sum of values with size [depth, nseqv, ...]

alpha: attention factors of size [nseqv, nseqq, ...]

Utils for array manipulation

NNHelferlein.crop_array — Function

function crop_array(x, crop_sizes)

Crop a n-dimensional array to the given size. Cropping is always centered (i.e. a margin is removed).

Arguments:

x: n-dim AbstractArray

crop_sizes: Tuple of target sizes to which the array is cropped. Allowed values are Int or :. If crop_sizes defines less dims as x has, the remaining dims will not be cropped (assuming :). If a demanded crop size is bigger as the actual size of x, it is ignored.

NNHelferlein.blowup_array — Function

function blowup_array(x, n)

Blow up an array x with an additional dimension and repeat the content of the array n times.

Arguments:

x: Array of any dimension

n: number of repeats. ´n == 1´ will return an

array with an addritional dimension of size 1.

Examples:

julia> x = [1,2,3,4]; blowup_array(x, 3)
4×3 Array{Int64,2}:
 1  1  1
 2  2  2
 3  3  3
 4  4  4

julia> x = [1 2; 3 4]; blowup_array(x, 3)
2×2×3 Array{Int64,3}:
[:, :, 1] =
 1  2
 3  4

[:, :, 2] =
 1  2
 3  4

[:, :, 3] =
 1  2
 3  4

NNHelferlein.recycle_array — Function

function recycle_array(x, n)

Recycle an array x along the last dimension and repeat the content of the array n times. The number of dims stays unchanged, but the array valueas are repeated n times.

Arguments:

x: Array of any dimension

n: number of repeats. ´n == 1´ will return an unchanged array.

Examples:

julia> recycle_array([1,2],3)
6-element Array{Int64,1}:
 1
 2
 1
 2
 1
 2

julia> x = [1 2; 3 4]
2×2 Array{Int64,2}:
 1  2
 3  4

julia> recycle_array(x,3)
2×6 Array{Int64,2}:
 1  2  1  2  1  2
 3  4  3  4  3  4

NNHelferlein.de_embed — Function

function de_embed(x; remove_dim=true)

Replace the maximum of the first dimension of an n-dimensional array by its index (aka argmax()). If remove_dim=false the first dim is preserved with size=1; otherwise the returned array has the first dimension removed.

Utils for fixing types in GPU context

NNHelferlein.init0 — Function

function init0(siz...)

Initialise a vector or array of size siz with zeros. If a GPU is detected type of the returned value is KnetArray{Float32}, otherwise Array{Float32}.

Examples:

julia> init0(2,10)
2×10 Array{Float32,2}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

 julia> init0(0,10)
 0×10 Array{Float32,2}

NNHelferlein.convert2KnetArray — Function

function convert2KnetArray(x)

Convert an array x to a KnetArray{Float32} or KnetArray{Int32}only in GPU context (ifCUDA.functional()) or to anArray{Float32}/Array{Int32}` otherwise.