Chains

NNHelferlein.NeuNet — Type

abstract type NeuNet end

Mother type for DNN hierarchy with implementation for a chain of layers.

Signatures:

(n::NeuNet)(x) = (for l in n.layers; x = l(x); end; x)
(m::NeuNet)(d::Knet.Data) = mean( m(x,y) for (x,y) in d)

NNHelferlein.Classifier — Type

struct Classifier <: NeuNet

Classifyer with nll loss.

Signatures:

(m::Classifier)(x,y) = nll(m(x), y)

NNHelferlein.Regressor — Type

struct Regressor

Regression network with square loss.

Signatures:

(m::Regression)(x,y) = sum(abs2.( m(x) - y))

NNHelferlein.Chain — Type

struct Chain

Simple wrapper tu chain layers afer each other.

Layers

Missing docstring.

Missing docstring for Layer. Check Documenter's build log for details.

Fully connected layers

NNHelferlein.Dense — Type

struct Dense  <: Layer

Default Dense layer.

Constructors:

Dense(w, b, actf): default constructor

Dense(i::Int, j::Int; actf=sigm): layer of j neurons with i inputs.

Dense(h5::HDF5.File, group::String; trainable=false, actf=sigm):

Dense(h5::HDF5.File, kernel::String, bias::String; trainable=false, actf=sigm): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

NNHelferlein.Embed — Type

struct Embed <: Layer

Simple type for an embedding layer to embed a virtual onehot-vector into a smaller number of neurons by linear combination. The onehot-vector is virtual, because not the vector, but only the index of the "one" in the vector has to be provided as Integer value (or a minibatch of integers).

Fields:

w

actf

Constructors:

Embed(i,j; actf=identity) = new(param(j,i), actf): with input size i, output size j and default activation function idendity.

Signatures:

(l::Embed)(x) = l.actf.(w[:,x]) default embedding of input vector x.

NNHelferlein.Predictions — Type

struct Predictions <: Layer

Simple wrapper around a Dense layer without activation function that can used as output layer (because all loss-functions of the package assume raw output activations).

Constructors:

Predictions(i::Int, j:Int): with input size i, output size j activation function idendity.

Predictions(h5::HDF5.File, group::String; trainable=false):

Predictions(h5::HDF5.File, kernel::String, bias::String; trainable=false): with an hdf5-object and group name of the output layer.

Convolutional

NNHelferlein.Conv — Type

struct Conv  <: Layer

Default Conv layer.

Constructors:

Conv(w, b, padding, actf): default constructor

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu; kwargs...): layer with o kernels of size (w1,w2) for an input of i layers.

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu):

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

Keyword arguments:

padding=0: the number of extra zeros implicitly concatenated at the start and end of each dimension.

stride=1: the number of elements to slide to reach the next filtering window.

dilation=1: dilation factor for each dimension.

... See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function conv4() are supported.

NNHelferlein.DeConv — Type

struct DeConv  <: Layer

Default deconvolution layer.

Constructors:

DeConv(w, b, actf, kwargs...): default constructor

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu, kwargs...): layer with o kernels of size (w1,w2) for an input of i channels.

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu):

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu): layer imported from a hdf5-file from tensorflow with the hdf-object hdfo and the group name group.

Keyword arguments:

padding=0: the number of extra zeros implicitly concatenated at the start and end of each dimension (applied to the output).

stride=1: the number of elements to slide to reach the next filtering window (applied to the output).

... See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function deconv4() are supported.

Missing docstring.

Missing docstring for Pool. Check Documenter's build log for details.

NNHelferlein.UnPool — Type

struct UnPool <: Layer

Unpooling layer.

Constructors:

UnPool(;kwargs...): user-defined unpooling

Recurrent

Missing docstring.

Missing docstring for RSeqClassifier. Check Documenter's build log for details.

Others

NNHelferlein.Flat — Type

struct Flat <: Layer

Default flatten layer.

Constructors:

Flat(): with no options.

NNHelferlein.PyFlat — Type

struct PyFlat <: Layer

Flatten layer with optional Python-stype flattening (row-major). This layer can be used if pre-trained weight matrices from tensorflow are applied after the flatten layer.

Constructors:

PyFlat(; python=true): if true, row-major flatten is performed.

NNHelferlein.Softmax — Type

struct Softmax <: Layer

Simple softmax layer to compute softmax probabilities.

Constructors:

Softmax()

NNHelferlein.Dropout — Type

struct Dropout <: Layer

Dropout layer. Implemented with help of Knet's dropout() function that evaluates AutoGrad.recording() to detect if in training or inprediction. Dropouts are applied only if prediction.

Constructors:

Dropout(p) with the dropout rate p.

NNHelferlein.BatchNorm — Type

struct BatchNorm <: Layer

Batchnormalisation layer. Implemented with help of Knet's batchnorm() function that evaluates AutoGrad.recording() to detect if in training or in prediction. In training the moments are updated to record the running averages; in prediction the moments are applied, but not modified.

Constructors:

Batchnom() will initialise the moments with Knet.bnmoments().

Data providers

NNHelferlein.DataLoader — Type

abstract type DataLoader

Mother type for minimatch iterators.

Tabular data

Tabular data is normally provided in table form (csv, ods) row-wise, i.e. one sample per row. The helper functions can read the tables and generate Knet compatible iterators of minibatches.

NNHelferlein.dataframe_read — Function

dataframe_read(fname)

Read a data table from an CSV-file with one sample per row and return a DataFrame with the data. (ODS-support is removed because of frequent PyCall compatibility issues of the OdsIO package).

NNHelferlein.dataframe_minibatches — Function

dataframe_minibatches(data::DataFrames.DataFrame; size=256, ignore=[], teaching=:y, regression=true)

Make Knet-conform minibatches from a dataframe with one sample per row.

ignore: defines a list of column names to be ignored

teaching: defines the column with teaching input. Default is ":y"

regression: if true, the teaching input is interpreted as scalar value (converted to Float32); otherwise teaching input is used as class labels (converted to UInt8).

NNHelferlein.dataframe_split — Function

function dataframe_split(df::DataFrames.DataFrame; teaching=:y, fr=0.2, balanced=true)

Split data, organised row-wise in a DataFrame into train and valid sets.

Arguments:

df: data

teaching: name or index of column with teaching input (y)

fr: fraction of data to be used for validation

balanced: if true, result datasets will be balanced by oversampling. Returned datasets will be bigger as expected but include the same numbers of samples for each class.

Image data

Images as data should be provided in directories with the directory names denoting the class labels. The helpers read from the root of a directory tree in which the first level of sub-dirs tell the class label. All images in the tree under a class label are read as instances of the respective class. The following tree will generate the classes daisy, rose and tulip:

image_dir/
├── daisy
│   ├── 01
│   │   ├── 01
│   │   ├── 02
│   │   └── 03
│   ├── 02
│   │   ├── 01
│   │   └── 02
│   └── others
├── rose
│   ├── big
│   └── small
└── tulip

NNHelferlein.ImageLoader — Type

struct ImageLoader <: DataLoader
    dir
    i_paths
    i_classes
    classes
    batchsize
    shuffle
    train
    aug_pipl
    pre_proc
    pre_load
    i_images
end

Iterable image loader to provide minibatches of images as 4-d-arrays (x,y,rgb,mb).

NNHelferlein.mk_image_minibatch — Function

function mk_image_minibatch(dir, batchsize; split=false, fr=0.2,
                            balanced=false, shuffle=true, train=true,
                            aug_pipl=nothing, pre_proc=nothing)

Return one or two iterable image-loader-objects that provides minibatches of images. For training each minibatch is a tupel (x,y) with x: 4-d-array with the minibatch of data and y: vector of class IDs as Int.

Arguments:

dir: base-directory of the image dataset. The first level of sub-dirs are used as class names.

batchsize: size of minibatches

Keyword arguments:

split: return two iterators for training and validation

fr: split fraction

balanced: return balanced data (i.e. same number of instances for all classes). Balancing is achieved via oversampling

shuffle: if true, shuffle the images everytime the iterator restarts

train: if true, minibatches with (x,y) Tuples are provided, if false only x (for prediction)

pre_load: if true all images are loaded in advance; otherwise images are loaded on demand durng training. (option is not implemented yet!)

aug_pipl: augmentation pipeline for Augmentor.jl. Augmentation is performed before the pre_proc-function is applied

pre_proc: function with preprocessing and augmentation algoritms of type x = f(x). In contrast to the augmentation that modifies images, is pre_proc working on Arrays{Float32}.

pre_load=false: read all images from disk once when populating the loader (requires loads of memory, but speeds up training).

NNHelferlein.get_class_labels — Function

function get_class_labels(d::DataLoader)

Extracts a list of class labels from a DataLoader.

NNHelferlein.image2array — Function

function image2array(img)

Take an image and return a 3d-array for RGB and a 2d-array for grayscale images with the colour channels as last dimension.

NNHelferlein.array2image — Function

function array2image(arr)

Take a 3d-array with colour channels as last dimension or a 2d-array and return an array of RGB or of Gray as Image.

NNHelferlein.array2RGB — Function

function array2RGB(arr)

Take a 3d-array with colour channels as last dimension or a 2d-array and return alwasy an array of RGB as Image.

Training

NNHelferlein.tb_train! — Function

function tb_train!(mdl, opti, trn, vld=nothing; epochs=1,
                  lr_decay=nothing, lrd_freq=1, l2=0.0,
                  eval_size=0.2, eval_freq=1,
                  acc_fun=nothing,
                  mb_loss_freq=100,
                  cp_freq=nothing, cp_dir="checkpoints",
                  tb_dir="logs", tb_name="run",
                  tb_text="""Description of tb_train!() run.""",
                  opti_args...)

Train function with TensorBoard integration. TB logs are written with the TensorBoardLogger.jl package. The model is updated (in-place) and the trained model is returned.

Arguments:

mdl: model; i.e. forward-function for the net

opti: Knet-stype optimiser iterator

trn: training data; iterator to provide (x,y)-tuples with minibatches

vld: validation data; iterator to provide (x,y)-tuples with minibatches. Set to nothing, if not defined.

Keyword arguments:

Optimiser:

epochs=1: number of epochs to train

lr_decay=nothing: Leraning rate decay if not nothing: factor (<1) to reduce the lr every epoch as lr *= lr_decay.

lrd_freq=1: frequency of learning rate decay steps. Default is to modify the lr after every epoch

l2=0.0: L2 regularisation; implemented as weight decay per parameter

opti_args...: optional keyword arguments for the optimiser can be specified (i.e. lr, gamma, ...).

Model evaluation:

eval_size=0.2: fraction of validation data to be used for calculating loss and accuracy for train and validation data during training.

eval_freq=1: frequency of evaluation; default=1 means evaluation is calculated after each epoch. With eval_freq=10 eveluation is calculated 10 times per epoch.

acc_fun=nothing: function to calculate accuracy. The function is called with 2 arguments: fun(predictions, teaching) where predictions is the output of a model call and a matrix and teaching is the teaching input (y). For classification tasks, accuracy from the Knet package is a good choice. For regression a correlation or mean error may be used (i.e. acc_fun=(x,y)->sum(abs, x.-y)).

mb_loss_freq=100: frequency of training loss reporting. default=100 means that 100 loss-values per epoch will be logged to TensorBoard. If mblossfreq is greater then the number of minibatches, loss is logged for each minibatch.

cp_freq=nothing: frequency of model checkpoints written to disk. Default is nothing, i.e. no checkpoints are written. To write the model after each epoch with name model use freq=1; to write every 2 epochs freq=0.5.

cp_dir="checkpoints": directory for checkpoints

TensorBoard:

TensorBoard log-directory is created from 3 parts: tb_dir/tb_name/<current date time>.

tb_dir="logs": root directory for tensorborad logs.

tb_name="run": name of training run. tb_name will be used as directory name and should not include whitespace

tb_text: description to be included in the TensorBoard log as text log.

Evaluation

NNHelferlein.predict — Function

function predict(mdl, x; softmax=false)

Return the prediction for x.

Arguments:

mdl: executable network model x: tensor, minibatch or iterator providing minibatches of input data softmax: if true and if model is a ::Classifier the prediction softmax probabilities are rezrned instead of raw activations.

NNHelferlein.predict_top5 — Function

function predict_top5(mdl, x; top_n=5, classes=nothing)

Run the model mdl for data in x and prints the top 5 predictions as softmax probabilities.

Arguments:

top_n: print top n hits instead of 5 classes may be a list of human readable class labels.

ImageNet tools

NNHelferlein.preproc_imagenet — Function

function preproc_imagenet(img)

Image preprocessing for pre-trained ImageNet examples. Preprocessing includes

bring RGB colour values into a range 0-255

standardise of colour values by substracting mean colour values (103.939, 116.779, 123.68) from RGB

changing colour channel sequence from RGB to BGR

Resize is not done, because this may be part of the augmentation pipeline.

Examples:

The function can be used with the image loader; for prediction with a trained model as:

pipl = CropRatio(ratio=1.0) |> Resize(224,224)
images = mk_image_minibatch("./example_pics", 16;
                    shuffle=false, train=false,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

And for training something like:

pipl = Either(1=>FlipX(), 1=>FlipY(), 2=>NoOp()) |>
       Rotate(-5:5) |>
       ShearX(-5:5) * ShearY(-5:5) |>
       RCropSize(224,224)

dtrn, dvld = mk_image_minibatch("./example_pics", 16;
                    split=true, fr=0.2, balanced=false,
                    shuffle=true, train=true,
                    aug_pipl=pipl,
                    pre_proc=preproc_imagenet)

NNHelferlein.predict_imagenet — Function

function predict_imagenet(mdl, x; top_n=5)

Predict the ImageNet-class of images from the predefined list of class labels.

Other utils

NNHelferlein.crop_array — Function

function crop_array(x, crop_sizes)

Crop a n-dimensional array to the given size. Cropping is always centered (i.e. a margin is removed).

Arguments:

x: n-dim AbstractArray

crop_sizes: Tuple of target sizes to which the array is cropped. Allowed values are Int or :. If crop_sizes defines less dims as x has, the remaining dims will not be cropped (assuming :). If a demanded crop size is bigger as the actual size of x, it is ignored.

NNHelferlein.init0 — Function

function init0(siz)

Initialise a vector or array of size siz with zeros. If a GPU is detected type of the returned value is KnetArray{Float32}, otherwise Array{Float32}.