API doc of all exported functions are listed here:

Chains

NNHelferlein.AbstractNN — Type

abstract type AbstractNN

Mother type for AbstractNN hierarchy with implementation for a chain of layers.

Signatures:

(m::AbstractNN)(x): run the AbstractArray x througth all layers and return the output

(m::AbstractNN)(x,y): Calculate the loss for one minibatch x and teaching input y

(m::AbstractNN)(d::Knet.Data): Calculate the loss for all minibatches in d

(m::AbstractNN)(d::Tuple): Calculate the loss for all minibatches in d

(m::AbstractNN)(d::NNHelferlein.DataLoader): Calculate the loss for all minibatches in d if teaching input is included (i.e. elements of d are tuples). Otherwise return the out of all minibatches as one array with samples as columns.

```

NNHelferlein.AbstractChain — Type

abstract type AbstractChain

Mother type for AbstractChain hierarchy with implementation for a chain of layers. By default every AbstractChain has a property layers with a iterable list of AbstractLayers or AbstractChains that are executed recursively.

Non-standard Chains in which Layers are not execueted sequnetially (such as ResnetBlocks) must provide a custom implementation with the signature chain(x).

Signatures:

(m::AbstractChain)(x): run the AbstractArray x througth all layers and return the output

```

NNHelferlein.add_layer! — Function

function add_layer!(n::Union{NNHelferlein.AbstractNN, NNHelferlein.AbstractChain}, l)

Add a layer l or a chain to a model n. The layer is always added at the end of the chains. The modified model is returned.

Base.:+ — Function

function +(n::Union{NNHelferlein.AbstractNN, NNHelferlein.AbstractChain}, l::Union{AbstractLayer, AbstractChain})
function +(l1::AbstractLayer, l2::Union{AbstractLayer, AbstractChain})

The plus-operator is overloaded to be able to add layers and chains to a network.

The second form returns a new chain if 2 Layers are added.

Example:

julia> mdl = Classifier() + Dense(2,5)
julia> print_network(mdl)

NNHelferlein neural network summary:
Classifier with 1 layers,                                           15 params
Details:
 
    Dense layer 2 → 5 with sigm,                                    15 params
 
Total number of layers: 1
Total number of parameters: 15

julia> mdl = mdl + Dense(5,5) + Dense(5,1, actf=identity)
julia> print_network(mdl)

NNHelferlein neural network summary:
Classifier with 3 layers,                                           51 params
Details:
 
    Dense layer 2 → 5 with sigm,                                    15 params
    Dense layer 5 → 5 with sigm,                                    30 params
    Dense layer 5 → 1 with identity,                                 6 params
 
Total number of layers: 3
Total number of parameters: 51

Missing docstring.

Missing docstring for print_network. Check Documenter's build log for details.

NNHelferlein.Classifier — Type

struct Classifier <: AbstractNN

Classifier with default nll loss. An alternative loss function can be supplied as keyword argument. The function must provide a signature to be called as loss(model(x), y).

Constructors:

Classifier(layers...; loss=Knet.nll)

Signatures:

(m::Classifier)(x,y) = m.loss(m(x), y)

NNHelferlein.Regressor — Type

struct Regressor <: AbstractNN

Regression network with square loss as loss function.

Constructors:

Regressor(layers...; loss=mean_squared_error.nll)

Signatures:

(m::Regression)(x,y) = mean(abs2, Array(m(x)) - y)

Missing docstring.

Missing docstring for Transformer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TokenTransformer. Check Documenter's build log for details.

NNHelferlein.Chain — Type

struct Chain <: AbstractChain

Simple wrapper to chain layers and execute them one after another.

NNHelferlein.VAE — Type

struct VAE   <: AbstractNN

Type for a generic variational autoencoder.

Constructor:

VAE(encoder, decoder)

Separate predefind chains (ideally, but not necessarily of type Chain) for encoder and decoder must be specified. The VAE needs the 2 parameters mean and variance to define the distribution of each code-neuron in the bottleneck-layer. In consequence the encoder output must be 2 times the size of the decoder input (in case of dense layers: if encoder output is a 8-value vector, 4 codes are defined and the decoder input is a 4-value vector; in case of convolutional layers the number of encoder output channels must be 2 times the number of the encoder input channels - see the examples).

Signatures:

(vae::VAE)(x)
(vae::VAE)(x,y)

Called with one argument, predict will be executed; with two arguments (args x and y should be identical for the autoencoder) the loss will be returned.

Details:

The loss is calculated as the sum of element-wise error squares plus the Kullback-Leibler-Divergence to adapt the distributions of the bottleneck codes:

\[\mathcal{L} = \frac{1}{2} \sum_{i=1}^{n_{outputs}} (t_{i}-o_{i})^{2} - \frac{1}{2} \sum_{j=1}^{n_{codes}}(1 + ln\sigma_{c_j}^{2}-\mu_{c_j}^{2}-\sigma_{c_j}^{2}) \]

Output of the autoencoder is cropped to the size of input before loss calculation (and before prediction); i.e. the output has always the same dimensions as the input, even if the last layer generates a bigger shape.

KL-training parameters:

The parameter β is by default set to 1.0, i.e. mean-squared error and KL has the same weights. The functions set_beta(vae, beta) and get_beta(vae) can be used to set and get the β used in training. With β=0.0 no KL-loss will be used.

NNHelferlein.get_beta — Function

function get_beta(vae::VAE; ramp=false)

Return a Dict with the current VAE-parameters beta and ramp-up.

Arguments:

ramp=false: if true, a vector of β for all ramp-up steps is returned. This way, the ramp-up phase can be visualised: <img src="./assets/vae-beta-range.png"/>

NNHelferlein.set_beta! — Function

function setbeta!(vae::VAE, βmax; ramp_up=false, steps=0)

Helper to set the current value of the VAE-parameter beta and ramp-up settings.

VAE loss is calculated as (mean of error squares) + β * (mean of KL divergence).

Ramp-up:

In case of ramp_up=true, β starts with almost 0.0 (sigm(-10.0) ≈4.5e-5) and reaches almost 1.0 after steps steps, following a sigmoid curve. steps should be more than 25, to avoid rounding errors in the calculation of the derivative of the sigmoid function.

Layers

NNHelferlein.AbstractLayer — Type

abstract type AbstractLayer
abstract type Layer

Mother type for layers hierarchy. (The type Layer is kept for backward compatibility)

Fully connected layers

NNHelferlein.Dense — Type

struct Dense  <: AbstractLayer

Default Dense layer.

Constructors:

Dense(w, b, actf): default constructor, w are the weights and b the bias.

Dense(i::Int, j::Int; actf=sigm, init=..): layer of j neurons with i inputs. Initialiser is xavieruniform for actf=sigm and xaviewnormal otherwise.

Dense(h5::HDF5.File, group::String; trainable=false, actf=sigm): kernel and bias are loaded by the specified group.

Dense(h5::HDF5.File, kernel::String, bias::String; trainable=false, actf=sigm): layer imported from a hdf5-file from TensorFlow with the hdf-object h5 and the group name group.

Missing docstring.

Missing docstring for Linear. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Embed. Check Documenter's build log for details.

Convolutional

NNHelferlein.Conv — Type

struct Conv  <: AbstractLayer

Default Conv layer.

Constructors:

Conv(w1::Int, w2::Int, i::Int, o::Int; actf=relu; kwargs...): layer with o kernels of size (w1,w2) for an input of i channels.

Conv(w1::Int, w2::Int, w3::Int, i::Int, o::Int; actf=relu; kwargs...): layer with 3-dimensional kernels for 3D convolution (requires 5-dimensional input)

Conv(w1::Int, i::Int, o::Int; actf=relu; kwargs...): layer with o kernels of size (1,w1) for an input of i channels. This 1-dimensional convolution uses a 2-dimensional kernel with a first dimension of size 1. Input and output contain an empty firfst dimension of size 1. If padding, stride or dilation are specified, 2-tuples must be specified to correspond with the 2-dimensional kernel (e.g. padding=(0,1) for a 1-padding along the 1D sequence).

Constructors to read parameters from Tensorflow/Keras HDF-files:

Conv(h5::HDF5.File, kernel::String, bias::String; trainable=false, actf=Knet.relu, use_bias=true, kwargs...): Import parameters from HDF file h5 with kernel and bias specifying the full path to weights and biases, respectively.

Conv(h5::HDF5.File, group::String; trainable=false, actf=relu, tf=true, use_bias=true): Import a conv-layer from a default TF/Keras HDF5 file. If tf=false, group defines the full path to the parameters group/kernel:0 and group/bias:0. If tf=true, group defines the only the group name and parameters are addressed as model_weights/group/group/kernel:0 and model_weights/group/group/bias:0.

Keyword arguments:

padding=0: the number of extra zeros implicitly concatenated at the start and end of each dimension.

stride=1: the number of elements to slide to reach the next filtering window.

dilation=1: dilation factor for each dimension.

... See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function conv4() are supported.

Missing docstring.

Missing docstring for DeConv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ResNetBlock. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DepthwiseConv. Check Documenter's build log for details.

NNHelferlein.Pool — Type

struct Pool <: AbstractLayer

Pooling layer.

Constructors:

Pool(;kwargs...): max pooling; without kwargs, 2-pooling is performed.

Keyword arguments:

window=2: pooling window size (same for all directions)

...: See the Knet documentation for Details: https://denizyuret.github.io/Knet.jl/latest/reference/#Convolution-and-Pooling. All keywords to the Knet function pool are supported.

Missing docstring.

Missing docstring for UnPool. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pad. Check Documenter's build log for details.

Recurrent

NNHelferlein.RecurrentUnit — Type

abstract type RecurrentUnit end

Supertype for all recurrent unit types. Self-defined recurrent units which are a child of RecurrentUnit can be used inside the 'Recurrent' layer.

Interface

All subtypes of RecurrentUnit must provide the followning:

a constructor with signature Type(n_inputs, n_units; kwargs) and arbitrary keyword arguments.

an implementation of signature (o::Recurrent)(x) where x is a 3d- or 2d-array of shape [fan-in, mb-size, 1] or [fan-in, mb-size]. The function must return the result of one forward computation for one step and return the hidden state and set the internal fields h and optionally c.

a field h (to store the last hidden state)

an optional field c, if the cell state is to be stored such as in a lstm unit.

Missing docstring.

Missing docstring for Recurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_hidden_states. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_cell_states. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_hidden_states!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_cell_states!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reset_hidden_states!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reset_cell_states!. Check Documenter's build log for details.

Transformers

Missing docstring.

Missing docstring for TFEncoder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TFEncoderLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TFDecoder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TFDecoderLayer. Check Documenter's build log for details.

These layers are used by the Transformer and TokenTransformer types to build Bert-like transformer networks.

Others

NNHelferlein.Flat — Type

struct Flat <: AbstractLayer

Default flatten layer.

Constructors:

Flat(): with no options.

NNHelferlein.flatten — Function

flatten(x)

Flatten a tensor to a matrix, preserving the last dimension.

NNHelferlein.PyFlat — Type

struct PyFlat <: AbstractLayer

Flatten layer with optional Python-stype flattening (row-major). This layer can be used if pre-trained weight matrices from tensorflow are applied after the flatten layer.

Constructors:

PyFlat(; python=true): if true, row-major flatten is performed.

Missing docstring.

Missing docstring for FeatureSelection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Activation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Softmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Logistic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Dropout. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BatchNorm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LayerNorm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaussianNoise. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GlobalAveragePooling. Check Documenter's build log for details.

Missing docstring.

Missing docstring for global_average_pooling. Check Documenter's build log for details.

Attention Mechanisms

Missing docstring.

Missing docstring for AttentionMechanism. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttnBahdanau. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttnLuong. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttnDot. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttnLocation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttnInFeed. Check Documenter's build log for details.

Data providers

NNHelferlein.DataLoader — Type

abstract type DataLoader

Mother type for minibatch iterators.

Iteration utilities

NNHelferlein.PartialIterator — Type

struct PartialIterator <: DataLoader

The PartialIterator wraps any iterator and will only iterate the states specified in the list indices.

Constuctors

PartialIterator(inner, indices; shuffle=true)

Type of the states must match the states of the wrapped iterator inner. A nothing element may be given to specify the first iterator element.

If shuffle==true, the list of indices are shuffled every time the PartialIterator is started.

NNHelferlein.split_minibatches — Function

function split_minibatches(it, at=0.8; shuffle=true)

Return 2 iterators of type PartialIterator which iterate only parts of the states of the iterator it. Be aware that the partial iterators will not contain copies of the data but instead forward the data provided by the iterator it.

The function can be used to split an iterator of minibatches into train- and validation iterators, without copying any data. As the PartialIterator objects work with the states of the inner iterator, it is important not to shuffle the inner iterator (in this case the composition of the partial iterators would change and training and validation data may be mixed!).

Arguments:

it: Iterator to be splitted. The list of allowed states is created by performing a full iteration once.

at: Split point. The first returned iterator will include the given fraction (default: 80%) of the states.

shuffle: If true, the elements are shuffled at each restart of the iterator.

Missing docstring.

Missing docstring for MBNoiser. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MBMasquerade. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUIterator. Check Documenter's build log for details.

Tabular data

Tabular data is normally provided in table form (csv, ods) row-wise, i.e. one sample per row. The helper functions can read the tables and generate Knet compatible iterators of minibatches.

Missing docstring.

Missing docstring for dataframe_read. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe_minibatch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe_split. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mk_class_ids. Check Documenter's build log for details.

Image data

Images as data should be provided in directories with the directory names denoting the class labels. The helpers read from the root of a directory tree in which the first level of sub-dirs tell the class label. All images in the tree under a class label are read as instances of the respective class. The following tree will generate the classes daisy, rose and tulip:

image_dir/
├── daisy
│   ├── 01
│   │   ├── 01
│   │   ├── 02
│   │   └── 03
│   ├── 02
│   │   ├── 01
│   │   └── 02
│   └── others
├── rose
│   ├── big
│   └── small
└── tulip

NNHelferlein.ImageLoader — Type

struct ImageLoader <: DataLoader
    dir
    i_paths
    i_classes
    classes
    batchsize
    shuffle
    train
    aug_pipl
    pre_proc
    pre_load
    i_images
end

Iterable image loader to provide minibatches of images as 4-d-arrays (x,y,rgb,mb).

Missing docstring.

Missing docstring for mk_image_minibatch. Check Documenter's build log for details.

NNHelferlein.get_class_labels — Function

function get_class_labels(d::DataLoader)

Extracts a list of class labels from a DataLoader.

Missing docstring.

Missing docstring for image2array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for array2image. Check Documenter's build log for details.

Missing docstring.

Missing docstring for array2RGB. Check Documenter's build log for details.

Text data

Missing docstring.

Missing docstring for WordTokenizer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_tatoeba_corpus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sequence_minibatch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pad_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for truncate_sequence. Check Documenter's build log for details.

Training

Missing docstring.

Missing docstring for tb_train!. Check Documenter's build log for details.

Evaluation and accuracy

Missing docstring.

Missing docstring for focal_nll. Check Documenter's build log for details.

Missing docstring.

Missing docstring for focal_bce. Check Documenter's build log for details.

Missing docstring.

Missing docstring for predict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for predict_top5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for minibatch_eval. Check Documenter's build log for details.

Missing docstring.

Missing docstring for squared_error_acc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for abs_error_acc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hamming_dist. Check Documenter's build log for details.

Missing docstring.

Missing docstring for peak_finder_acc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for confusion_matrix. Check Documenter's build log for details.

ImageNet tools

Missing docstring.

Missing docstring for preproc_imagenet_vgg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for preproc_imagenet_resnet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for preproc_imagenet_resnetv2, predict_imagenet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_imagenet_classes. Check Documenter's build log for details.

Other utils

Layers and helpers for transformers

Missing docstring.

Missing docstring for PositionalEncoding. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mk_padding_mask. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mk_peek_ahead_mask. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dot_prod_attn. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MultiHeadAttn. Check Documenter's build log for details.

Missing docstring.

Missing docstring for separate_heads. Check Documenter's build log for details.

Missing docstring.

Missing docstring for merge_heads. Check Documenter's build log for details.

Utils for array manipulation

Missing docstring.

Missing docstring for crop_array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for blowup_array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for recycle_array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for de_embed. Check Documenter's build log for details.

Utils for fixing types in GPU context

Missing docstring.

Missing docstring for init0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for convert2CuArray. Check Documenter's build log for details.

Missing docstring.

Missing docstring for convert2KnetArray. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ifgpu. Check Documenter's build log for details.

Missing docstring.

Missing docstring for emptyCuArray. Check Documenter's build log for details.

Utils for Bioinformatics

Missing docstring.

Missing docstring for aminoacid_tokenizer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for embed_blosum62. Check Documenter's build log for details.

Missing docstring.

Missing docstring for embed_vhse8. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EmbedAminoAcids. Check Documenter's build log for details.

Saving, loading and inspection of models

Missing docstring.

Missing docstring for save_network. Check Documenter's build log for details.

Missing docstring.

Missing docstring for load_network. Check Documenter's build log for details.

Missing docstring.

Missing docstring for copy_network. Check Documenter's build log for details.

Base.summary — Function

function summary(mdl)

Print a network summary of any model of Type AbstractNN, AbstractChain or AbstractLayer.

Missing docstring.

Missing docstring for print_network. Check Documenter's build log for details.

Datasets

Missing docstring.

Missing docstring for dataset_mit_nsr. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataset_mnist. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataset_fashion_mnist. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataset_iris. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataset_pfam. Check Documenter's build log for details.

Pretrained networks

Missing docstring.

Missing docstring for get_vgg16. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_resnet50v2. Check Documenter's build log for details.